knowledge_capsules/backend/requirements.txt
# --- Core Web Framework ---
fastapi
uvicorn[standard]

# --- Database (PostgreSQL ORM) ---
sqlalchemy
psycopg2-binary
alembic  # Для миграций базы данных

# --- Data Validation & Settings ---
pydantic
pydantic-settings

# --- Background Tasks ---
celery
redis

# --- Security & Authentication ---
passlib[bcrypt]  # For password hashing
python-jose[cryptography]  # For JWT tokens

# --- Email Service ---
sendgrid

# --- Environment Variables ---
python-dotenv

# --- CORS ---
# FastAPI includes CORS middleware, so no separate package is strictly needed
# unless more advanced features are required.

# --- Other Utilities ---
# greenlet might be needed by SQLAlchemy's async features
greenlet
